+++ games/phaser/diplomacy/src/client/game/game.ts
@@ -3,6 +3,7 @@ import { GameOver } from './scenes/GameOver';
 import { GamePlay } from './scenes/GamePlay';
 import { MainMenu } from './scenes/MainMenu';
 import { MyGames } from './scenes/MyGames';
+import { Tutorial } from './scenes/Tutorial';
 import * as Phaser from 'phaser';
 import { AUTO, Game } from 'phaser';
 import { Preloader } from './scenes/Preloader';
@@ -18,7 +19,7 @@ const config: Phaser.Types.Core.GameConfig = {
     width: '100%',
     height: '100%',
   },
-  scene: [Boot, Preloader, MainMenu, GamePlay, GameOver, MyGames],
+  scene: [Boot, Preloader, MainMenu, GamePlay, GameOver, MyGames, Tutorial],
 };
 
 const StartGame = (parent: string) => {